requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.30
